code/requirements.txt
ddsmoothing-python>=1.0.0
torch>=1.10.2
torchvision>=0.7.0
tqdm>=4.44.0
scipy>=1.5.0
statsmodels>=0.11.1
ancer-python==1.0.0 
joblib>=0.14.1
matplotlib>=3.3.4
numpy>=1.18.1
pandas>=1.1.5
scikit-learn>=0.22.1
autograd>=1.3
geomstats>=2.3.1 

